Spot Light

+++ Rapid/src/Graphics/PointLight.cpp
@@ -26,7 +26,9 @@ namespace Rapid {	namespace Graphics {
 		
 	}
 
-	void PointLight::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation, GLfloat diffuseIntensityLocation, GLfloat positionLocation, GLfloat constantLocation, GLfloat linearLocation, GLfloat exponentLocation)
+	void PointLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation, GLuint diffuseIntensityLocation,
+		GLuint positionLocation, GLuint constantLocation,
+		GLuint linearLocation, GLuint exponentLocation)
 	{
 		glUniform3f(ambientColorLocation, color.x, color.y, color.z);
 		glUniform1f(ambientIntensityLocation, ambientIntensity);
